Warning when variable is not used, __use__ macro and other
Added a new system whe na variable is not used there's an warning
but if the variable starts with '_' then warning will not show but
if you want to have your variable with out '_' at the start u can
use __use__ macro after it. Everything is shown ./examples/40-__use__.ba.